+++ tests/testthat/test_set_conditionalset.R
@@ -64,6 +64,6 @@ test_that("isSubset", {
 test_that("fields", {
   c <- ConditionalSet$new(function(x) TRUE)
   expect_equal(c$condition, function(x) TRUE)
-  expect_equal(c$class, list(x = UniversalSet$new()))
+  expect_equal(c$class, list(x = Universal$new()))
   expect_equal(c$elements, NA)
 })