+++ NEWS.md
@@ -1,3 +1,9 @@
+# set6 0.1.8.9000
+
+* UniversalSet renamed Universal, old class will be removed in v0.4.0.
+* LogicalSet renamed Logicals, old class will be removed in v0.4.0.
+* `Complex` now inherits from `Set`, incorrect methods for `isSubset, equals` have been removed.
+
 # set6 0.1.8
 
 * Patch for R-devel